Added mockups folder: initial screens for client and server.
Notes: The "Thesis" project can be directly imported to the WireframeSketcher tool's workspace (same for a corresponding eclipse plugin version). There is also a generated_as_img_screens folder, where the fore-mentioned screens are exported as image documents, for direct review (if one does not want to install this tool).